Delete refrences for a PyPi package in README.md
No need to upload a wheel to PyPi every time I make a new release when the user can just use `pip` and a URL to do the same thing.
It's not a python library, it is a CLI tool.

Maybe I can create a Continuous Deployment to do it for me, but no need.